% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svb_fit.R
\name{svb.fit}
\alias{svb.fit}
\title{Fit sparse variational Bayesian survival models.}
\usage{
svb.fit(
  Y,
  delta,
  X,
  lambda = 0.5,
  model = "partial",
  params = NULL,
  mu.init = NULL,
  s.init = NULL,
  g.init = NULL,
  maxiter = 1000,
  tol = 0.001,
  verbose = TRUE
)
}
\arguments{
\item{Y}{failure times}

\item{delta}{censoring indicator, 0: censored, 1: uncensored}

\item{X}{design matrix}

\item{lambda}{penalisation hyperparameter}

\item{model}{one of: \itemize{
    \item{\code{"partial"}} (defualt),
\item{\code{"exponential"}}
}}

\item{params}{additional hyperparameters, defualts: \itemize{
    \item{partial}: \code{c(a_0=1, b_0=ncol(X))},
    \item{exponential}: \code{c(a_0=1, b_0=ncol(X), a_omega=1, b_omega=1)}.
}}

\item{mu.init}{initial values for means, default: \code{rnorm(ncol(X))}}

\item{s.init}{initial values for standard deviations, default: \code{rep(0.2, ncol(X))}}

\item{g.init}{initial values for gamma, default: \code{rep(0.5, ncol(X))}}

\item{maxiter}{maximum number of iterations}

\item{tol}{convergence tolerance}

\item{verbose}{print additional information}
}
\description{
Fit sparse variational Bayesian survival models.
}
\examples{
n <- 200
omega <- 1
censoring_lvl <- 0.4

set.seed(1)
b <- c(1, 1, rep(0, 348))
p <- length(b)
X <- matrix(rnorm(n * p), nrow=n)
y <- runif(nrow(X))
Y <- log(1 - y) / - (exp(X \%*\% b) * omega)

delta  <- runif(n) > censoring_lvl   # 0: censored, 1: uncensored
Y[!delta] <- Y[!delta] * runif(sum(!delta))

lambda <- 0.5
f <- svb.fit(Y, delta, X)

}
