% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbo.R
\name{elbo}
\alias{elbo}
\title{Compute the evidence lower bound (ELBO) for the Cox model}
\usage{
elbo(Y, delta, X, fit, lambda, a0, b0, nrep = 10000)
}
\arguments{
\item{Y}{failure times}

\item{delta}{censoring indicator}

\item{X}{deisgn matrix}

\item{fit}{fit model}

\item{lambda}{model hyperparameter}

\item{a0}{model hyperparameter}

\item{b0}{model hyperparameter}

\item{nrep}{Monte Carlo samples}
}
\description{
Compute the evidence lower bound (ELBO) for the Cox model
}
