% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbo.R
\name{elbo}
\alias{elbo}
\title{Compute the evidence lower bound (ELBO) for the Cox model}
\usage{
elbo(Y, delta, X, fit, nrep = 10000, center = TRUE)
}
\arguments{
\item{Y}{failure times}

\item{delta}{censoring indicator}

\item{X}{deisgn matrix}

\item{fit}{fit model}

\item{nrep}{Monte Carlo samples}

\item{center}{center the design matrix}
}
\description{
Compute the evidence lower bound (ELBO) for the Cox model
}
