% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svb_fit.R
\name{svb.fit}
\alias{svb.fit}
\title{Fit sparse variational Bayesian proportional hazards model.}
\usage{
svb.fit(
  Y,
  delta,
  X,
  lambda = 0.5,
  a0 = 1,
  b0 = ncol(X),
  mu.init = NULL,
  s.init = rep(0.05, ncol(X)),
  g.init = rep(0.5, ncol(X)),
  maxiter = 1000,
  tol = 0.001,
  alpha = 1,
  center = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{Y}{Failure times}

\item{delta}{Censoring indicator, 0: censored, 1: uncensored}

\item{X}{Design matrix}

\item{lambda}{Penalisation parameter, default: \code{lambda=0.5}}

\item{a0}{Beta distribution parameter, default: \code{a0=1}.}

\item{b0}{Beta distribution parameter, default: \code{b0=rnorm(ncol(X))}.}

\item{mu.init}{Initial value for means, default taken from elasticnet fit}

\item{s.init}{initial values for standard deviations, default: \code{rep(0.05, ncol(X))}}

\item{g.init}{initial values for gamma, default: \code{rep(0.5, ncol(X))}}

\item{maxiter}{maximum number of iterations}

\item{tol}{convergence tolerance}

\item{alpha}{The elasticnet mixing parameter used for initialising \code{mu.init}, when \code{alpha=1} the lasso penalty is used and \code{alpha=0} the ridge penalty, values between 0 and 1 give a mixture of the two penalties.}

\item{center}{center X prior to fitting, increases numerical stability}

\item{verbose}{print additional information}
}
\description{
Fit sparse variational Bayesian proportional hazards model.
}
\examples{
n <- 200
p <- 1000
s <- 10
censoring_lvl <- 0.4

set.seed(1)
b <- sample(c(runif(s, -2, 2), rep(0, p-s)))
X <- matrix(rnorm(n * p), nrow=n)
Y <- log(1 - runif(n)) / -exp(X \%*\% b)

delta  <- runif(n) > censoring_lvl   # 0: censored, 1: uncensored
Y[!delta] <- Y[!delta] * runif(sum(!delta))

f <- svb.fit(Y, delta, X)

}
