% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svb_fit.R
\name{svb.fit}
\alias{svb.fit}
\title{Fit sparse variational Bayesian survival models.}
\usage{
svb.fit(
  Y,
  delta,
  X,
  lambda,
  params,
  mu.init = NULL,
  s.init = NULL,
  g.init = NULL,
  model = "partial",
  maxiter = 1000,
  tol = 0.001,
  verbose = TRUE
)
}
\arguments{
\item{Y}{observation times}

\item{delta}{Inidcator for right censoring 0: censored, 1: uncensored}

\item{X}{model matrix}

\item{lambda}{penalisation term}

\item{mu.init}{initial values for means}

\item{s.init}{initial values for standard deviations}

\item{g.init}{initial values for gamma, the inclusion probability}

\item{model}{one of "partial", "exponential"}

\item{maxiter}{maximum number of iterations}

\item{tol}{Stopping tolerance}

\item{verbose}{print information}
}
\description{
Fit sparse variational Bayesian survival models.
}
\examples{
n <- 250
omega <- 1
censoring_lvl <- 0.4

set.seed(1)
b <- c(1, 1, rep(0, 50))
p <- length(b)
X <- matrix(rnorm(n * p), nrow=n)
y <- runif(nrow(X))
Y <- log(1 - y) / - (exp(X \%*\% b) * omega)

delta  <- runif(n) > censoring_lvl   # 0: censored, 1: uncensored
Y[!delta] <- Y[!delta] * runif(sum(!delta))

lambda <- 0.5
svb.fit(Y, delta, X, lambda, c(1e-3, 1e-3))

}
